+++ CHANGELOG.md
@@ -6,6 +6,9 @@
 ### **Breaking** ⚠️
 - Moved `until` to be an option in `report`
 
+### Added
+- `report` option to treat category paths ("project/task") as their root ("project")
+
 ## [v0.6.0] - 2025-05-20
 ### **Breaking** ⚠️
 - Changed log file to be a CSV. To transition, use the following command: `sed 's/: /, /g' $FILE`.